Fix certMonitor labelSelector
