+++ internal/certmonitor/certmonitor.go
@@ -10,6 +10,8 @@ import (
 	"github.com/stackrox/acs-fleet-manager/fleetshard/pkg/fleetshardmetrics"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/tools/cache"
@@ -116,6 +118,9 @@ func (c *certMonitor) handleSecretCreation(obj interface{}) {
 	if !ok {
 		return
 	}
+	if !c.shouldProcessSecret(secret) {
+		return
+	}
 	c.processSecret(secret)
 }
 
@@ -140,6 +145,11 @@ func (c *certMonitor) handleSecretUpdate(oldObj, newObj interface{}) {
 			c.metrics.DeleteKeyCertMetric(newSecret.Namespace, newSecret.Name, oldKey)
 		}
 	}
+
+	if !c.shouldProcessSecret(newSecret) {
+		return
+	}
+
 	c.processSecret(newSecret)
 }
 
@@ -149,7 +159,65 @@ func (c *certMonitor) handleSecretDeletion(obj interface{}) {
 	if !ok {
 		return
 	}
+	if !c.shouldProcessSecret(secret) {
+		return
+	}
 	c.metrics.DeleteCertMetric(secret.Namespace, secret.Name)
+}
+
+func (c *certMonitor) shouldProcessSecret(s *corev1.Secret) bool {
+	for _, monitor := range c.config.Monitors {
+		if c.secretMatches(s, monitor) {
+			return true
+		}
+	}
+	return false
+}
+
+// secretMatches checks if a secret matches a monitor config
+func (c *certMonitor) secretMatches(s *corev1.Secret, monitor MonitorConfig) bool {
+	if s == nil {
+		return false
+	}
+	if len(monitor.Secret.Name) > 0 && s.Name != monitor.Secret.Name {
+		return false
+	}
+	if len(monitor.Namespace.Name) > 0 && s.Namespace != monitor.Namespace.Name {
+		return false
+	}
+	if monitor.Secret.LabelSelector != nil && !objectMatchesSelector(s, monitor.Secret.LabelSelector) {
+		return false
+	}
+
+	if monitor.Namespace.LabelSelector != nil {
+		ns, err := c.namespaceGetter.Get(s.Namespace)
+		if err != nil {
+			return false
+		}
+		if !objectMatchesSelector(ns, monitor.Namespace.LabelSelector) {
+			return false
+		}
+	}
+	return true
+}
+
+// objectMatchesSelector checks if object matches given label selector
+func objectMatchesSelector(obj runtime.Object, selector *metav1.LabelSelector) bool {
+	if selector == nil {
+		return true
+	}
+	labelselector, err := metav1.LabelSelectorAsSelector(selector)
+	if err != nil {
+		return false
+	}
+
+	metaObj, ok := obj.(metav1.Object)
+	if !ok {
+		return false
+	}
+
+	return labelselector.Matches(labels.Set(metaObj.GetLabels()))
+
 }
 
 // ValidateConfig checks the validity of Config
